// src/module/food/coupons/enum.ts
export enum CouponLevel {
  RESTAURANT = 'restaurant',
  GLOBAL = 'global',
}

export enum CouponType {
  FLAT = 'flat',
  UPTO = 'upto',
}

export enum DiscountSponseredBy {
  RESTAURANT = 'restaurant',
  BANK = 'bank',
}

export enum CouponCreatedBy {
  ADMIN = 'admin',
  VENDOR = 'vendor',
}

export enum CouponMappedBy {
  ADMIN = 'admin',
  VENDOR = 'vendor',
}

export enum CouponVendorMappingTimeLine {
  EXPIRED = 'expired',
  UPCOMING = 'upcoming',
  ACTIVE = 'active',
}